Move isNewer() into node.go
