+++ node_linux.go
@@ -7,8 +7,6 @@ import (
 	"strconv"
 	"syscall"
 	"time"
-
-	"github.com/restic/restic/debug"
 )
 
 func (node *Node) OpenForReading() (*os.File, error) {
@@ -70,31 +68,6 @@ func (node *Node) createFifoAt(path string) error {
 	return syscall.Mkfifo(path, 0600)
 }
 
-func (node *Node) isNewer(path string, fi os.FileInfo) bool {
-	if node.Type != "file" {
-		debug.Log("node.isNewer", "node %v is newer: not file", path)
-		return true
-	}
-
-	tpe := nodeTypeFromFileInfo(fi)
-	if node.Name != fi.Name() || node.Type != tpe {
-		debug.Log("node.isNewer", "node %v is newer: name or type changed", path)
-		return true
-	}
-
-	extendedStat := fi.Sys().(*syscall.Stat_t)
-	changeTime := time.Unix(extendedStat.Ctim.Unix())
-	inode := extendedStat.Ino
-	size := uint64(extendedStat.Size)
-
-	if node.ModTime != fi.ModTime() ||
-		node.ChangeTime != changeTime ||
-		node.Inode != inode ||
-		node.Size != size {
-		debug.Log("node.isNewer", "node %v is newer: timestamp, size or inode changed", path)
-		return true
-	}
-
-	debug.Log("node.isNewer", "node %v is not newer", path)
-	return false
+func changeTime(stat *syscall.Stat_t) time.Time {
+	return time.Unix(stat.Ctim.Unix())
 }
